desktop/requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
Pillow>=10.0.0
matplotlib>=3.8.0
PyPDF2>=3.0.0
pymupdf>=1.23.0
